check permissions for erase tools

+++ includes/admin/admin-functions.php
@@ -143,7 +143,9 @@ function zp_erase_natal_in_signs() {
 	if ( ! wp_verify_nonce( $_GET['_nonce'], 'zp_erase_natal_in_signs' ) ) {
 		return false;
 	}
-
+	if ( ! current_user_can( 'manage_zodiacpress_interps' ) ) {
+		return;
+	}
 	delete_option( 'zp_natal_planets_in_signs' );
 
 	$url = esc_url_raw( add_query_arg( array(
@@ -164,7 +166,9 @@ function zp_erase_natal_in_houses() {
 	if ( ! wp_verify_nonce( $_GET['_nonce'], 'zp_erase_natal_in_houses' ) ) {
 		return false;
 	}
-
+	if ( ! current_user_can( 'manage_zodiacpress_interps' ) ) {
+		return;
+	}
 	delete_option( 'zp_natal_planets_in_houses' );
 
 	$url = esc_url_raw( add_query_arg( array(
@@ -185,7 +189,9 @@ function zp_erase_natal_aspects() {
 	if ( ! wp_verify_nonce( $_GET['_nonce'], 'zp_erase_natal_aspects' ) ) {
 		return false;
 	}
-
+	if ( ! current_user_can( 'manage_zodiacpress_interps' ) ) {
+		return;
+	}
 	foreach ( zp_get_planets() as $planet ) {
 		$p = ( 'sun' == $planet['id'] ) ? 'main' : $planet['id'];
 		delete_option( 'zp_natal_aspects_' . $p );
